x3101-app/x3101-0210-backEnd/x0201-fastAPI/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
python-multipart
email-validator
python-dotenv
openpyxl
pandas
psutil
influxdb-client
docxtpl
